+++ CHANGELOG.md
@@ -22,8 +22,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added new `compas_timber._fabrication.JackRafterCutParams`.
 * Added new `compas_timber._fabrication.Drilling`.
 * Added new `compas_timber._fabrication.DrillingParams`.
+* Added new `compas_timber._fabrication.StepJoint`.
+* Added new `compas_timber._fabrication.StepJointNotch`.
+* Added new `compas_timber.connections.TStepJoint`.
+* Added new `utilities` module in `connections` package.
 
 ### Changed
+
+* Changed incorrect import of `compas.geometry.intersection_line_plane()` to `compas_timber.utils.intersection_line_plane()`
 
 * Added LayerName & ClearLayer Input to GH `BakeBoxMap` and `BakePlateBoxMap` Components.